refactor(racingGame): package 를 분리
