+++ src/main/java/racinggame/domain/Cars.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.domain;
 
 import com.google.common.base.Preconditions;
