+++ src/main/java/racinggame/domain/GameResult.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.domain;
 
 import java.util.Collections;
 import java.util.Map;
