+++ src/main/java/racinggame/domain/GameRound.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.domain;
 
 import com.google.common.base.Preconditions;
