+++ src/main/java/racinggame/domain/MoveDecider.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.domain;
 
 public interface MoveDecider {
     boolean canMove();
