+++ src/main/java/racinggame/domain/RacingGame.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.domain;
 
 public class RacingGame {
     private final GameResults gameResults;
