+++ src/main/java/racinggame/domain/RacingGameParameters.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.domain;
 
 public class RacingGameParameters {
     private final int gameRound;
