+++ src/main/java/racinggame/domain/RandomDecider.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.domain;
 
 import java.util.Random;
