+++ src/test/java/racinggame/domain/RacingGameParametersTest.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.domain;
 
 import org.junit.jupiter.api.Test;
