+++ src/test/java/racinggame/domain/WinnersTest.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.domain;
 
 import org.junit.jupiter.api.Test;
